+++ bin/metrics-es-node-graphite.rb
@@ -154,7 +154,7 @@ class ESNodeGraphiteMetrics < Sensu::Plugin::Metric::CLI::Graphite
       ].join('&')
     end
 
-    if Gem::Version.new(acquire_es_version) >= Gem::Version.new('1.0.0')
+    if es_version >= Gem::Version.new('1.0.0')
       stats = get_es_resource("/_nodes/_local/stats?#{stats_query_string}")
     else
       stats = get_es_resource("/_cluster/nodes/_local/stats?#{stats_query_string}")
@@ -166,13 +166,17 @@ class ESNodeGraphiteMetrics < Sensu::Plugin::Metric::CLI::Graphite
     metrics = {}
 
     if os_stat
-      metrics['os.load_average']                  = node['os']['load_average'][0]
-      metrics['os.load_average.1']                = node['os']['load_average'][0]
-      metrics['os.load_average.5']                = node['os']['load_average'][1]
-      metrics['os.load_average.15']               = node['os']['load_average'][2]
+      if es_version >= Gem::Version.new('2.0.0')
+        metrics['os.load_average']                  = node['os']['load_average']
+      else
+        metrics['os.load_average']                  = node['os']['load_average'][0]
+        metrics['os.load_average.1']                = node['os']['load_average'][0]
+        metrics['os.load_average.5']                = node['os']['load_average'][1]
+        metrics['os.load_average.15']               = node['os']['load_average'][2]
+      end
       metrics['os.mem.free_in_bytes']             = node['os']['mem']['free_in_bytes']
       # ... Process uptime in millis?
-      metrics['os.uptime']                        = node['os']['uptime_in_millis']
+      metrics['os.uptime']                        = node['os']['uptime_in_millis'] if node['os']['uptime_in_millis']
     end
 
     if process_stats
@@ -205,6 +209,7 @@ class ESNodeGraphiteMetrics < Sensu::Plugin::Metric::CLI::Graphite
 
       metrics['jvm.threads.count']                = node['jvm']['threads']['count']
       metrics['jvm.threads.peak_count']           = node['jvm']['threads']['peak_count']
+      metrics['jvm.uptime']                       = node['jvm']['uptime_in_millis']
     end
 
     node['indices'].each do |type, index|
@@ -226,18 +231,20 @@ class ESNodeGraphiteMetrics < Sensu::Plugin::Metric::CLI::Graphite
     metrics['http.current_open']                = node['http']['current_open']
     metrics['http.total_opened']                = node['http']['total_opened']
 
-    metrics['network.tcp.active_opens']         = node['network']['tcp']['active_opens']
-    metrics['network.tcp.passive_opens']        = node['network']['tcp']['passive_opens']
+    if node['network']
+      metrics['network.tcp.active_opens']         = node['network']['tcp']['active_opens']
+      metrics['network.tcp.passive_opens']        = node['network']['tcp']['passive_opens']
 
-    metrics['network.tcp.in_segs']              = node['network']['tcp']['in_segs']
-    metrics['network.tcp.out_segs']             = node['network']['tcp']['out_segs']
-    metrics['network.tcp.retrans_segs']         = node['network']['tcp']['retrans_segs']
-    metrics['network.tcp.attempt_fails']        = node['network']['tcp']['attempt_fails']
-    metrics['network.tcp.in_errs']              = node['network']['tcp']['in_errs']
-    metrics['network.tcp.out_rsts']             = node['network']['tcp']['out_rsts']
+      metrics['network.tcp.in_segs']              = node['network']['tcp']['in_segs']
+      metrics['network.tcp.out_segs']             = node['network']['tcp']['out_segs']
+      metrics['network.tcp.retrans_segs']         = node['network']['tcp']['retrans_segs']
+      metrics['network.tcp.attempt_fails']        = node['network']['tcp']['attempt_fails']
+      metrics['network.tcp.in_errs']              = node['network']['tcp']['in_errs']
+      metrics['network.tcp.out_rsts']             = node['network']['tcp']['out_rsts']
 
-    metrics['network.tcp.curr_estab']           = node['network']['tcp']['curr_estab']
-    metrics['network.tcp.estab_resets']         = node['network']['tcp']['estab_resets']
+      metrics['network.tcp.curr_estab']           = node['network']['tcp']['curr_estab']
+      metrics['network.tcp.estab_resets']         = node['network']['tcp']['estab_resets']
+    end
 
     if tp_stats
       node['thread_pool'].each do |pool, stat|